requirements.txt
 
streamlit 
langchain 
langchain-core 
langchain-google-genai 
langchain-community 
google-generativeai 
python-dotenv 
pypdf 
faiss-cpu 
tiktoken
numpy